backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pydantic
pydantic-settings
openai
slowapi
python-multipart
apscheduler
pytz
# DevIntel RAG dependencies
tiktoken
PyMuPDF
